Split functions into multiple files
So they can later be transfered to the protonats repository, so that
they can be shared amongst multiple protoc compiler plugins.
