+++ cmd/protoc-gen-go-nats/go.go
@@ -1,0 +1,5 @@
+package main
+
+import "strings"
+
+func unexport(s string) string { return strings.ToLower(s[:1]) + s[1:] }
